CasinoProyecto/requirements.txt
asgiref==3.10.0
Django==5.1.4
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
pillow==10.4.0
PyJWT==2.10.1
python-dotenv==1.0.0
qrcode==8.2
setuptools==80.9.0
sqlparse==0.5.3
